+++ htdma_code/controller/controller.py
@@ -1,3 +1,27 @@
+"""
+MIT License
+
+Copyright (c) 2023-25 Brian R. King
+
+Permission is hereby granted, free of charge, to any person obtaining a copy
+of this software and associated documentation files (the "Software"), to deal
+in the Software without restriction, including without limitation the rights
+to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+copies of the Software, and to permit persons to whom the Software is
+furnished to do so, subject to the following conditions:
+
+The above copyright notice and this permission notice shall be included in all
+copies or substantial portions of the Software.
+
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+SOFTWARE.
+
+"""
 
 from PySide2.QtWidgets import QFileDialog
 import PySide2.QtWidgets as Qw
@@ -6,6 +30,22 @@ from htdma_code.model.model import Model
 from htdma_code.view.main_window import MainWindow
 
 class Controller:
+    """
+    Handles the interaction logic between the model and the view components.
+
+    This class acts as the Controller in the Model-View-Controller (MVC) architectural pattern. It is responsible for
+    binding user actions or events from the GUI to appropriate model methods and keeping the view in sync with the
+    model's state.
+
+    Attributes
+    ----------
+    model : Model
+        The data model associated with the application.
+    main_view : MainWindow
+        The GUI main view of the application.
+    status_bar : QStatusBar
+        The status bar widget available in the application's main view.
+    """
     def __init__(self,model: Model,main_view: MainWindow):
         self.model = model
         self.main_view = main_view